My data structures are significantly slowing down SDL. Refactor (remove) them.
